requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
torch==2.1.1
transformers==4.36.2
huggingface-hub==0.19.4
pyyaml==6.0.1
pydantic==2.5.0
python-multipart==0.0.6
bitsandbytes==0.41.3